test sql agent with hints
