+++ tests/agent/sql_agent_llama.yaml
@@ -8,7 +8,7 @@ services:
     volumes:
       - ${TOOLSET_PATH}:/home/user/tools/ # tools
       - ${WORKDIR}/GenAIComps/comps:/home/user/comps
-      - ${WORKDIR}/TAG-Bench/:/home/user/TAG-Bench # SQL database
+      - ${WORKDIR}/TAG-Bench/:/home/user/TAG-Bench # SQL database and hints_file
     ports:
       - "9095:9095"
     ipc: host
@@ -17,7 +17,8 @@ services:
       strategy: sql_agent_llama
       db_name: ${db_name}
       db_path: ${db_path}
-      use_hints: false
+      use_hints: true
+      hints_file: /home/user/TAG-Bench/${db_name}_hints.csv
       recursion_limit: ${recursion_limit}
       llm_engine: vllm
       HUGGINGFACEHUB_API_TOKEN: ${HUGGINGFACEHUB_API_TOKEN}
@@ -26,11 +27,11 @@ services:
       temperature: ${temperature}
       max_new_tokens: ${max_new_tokens}
       streaming: false
-      tools: /home/user/tools/custom_tools.yaml
+      tools: /home/user/tools/sql_agent_tools.yaml # change back to custom_tools.yaml
       require_human_feedback: false
       no_proxy: ${no_proxy}
       http_proxy: ${http_proxy}
       https_proxy: ${https_proxy}
       port: 9095
-      GOOGLE_CSE_ID: ${GOOGLE_CSE_ID}
-      GOOGLE_API_KEY: ${GOOGLE_API_KEY}
+      GOOGLE_CSE_ID: ${GOOGLE_CSE_ID} #delete
+      GOOGLE_API_KEY: ${GOOGLE_API_KEY} # delete
